Endre mentorAntallTimer til Double
og decimal i postgres

+++ src/main/java/no/nav/tag/tiltaksgjennomforing/avtale/AvtaleInnhold.java
@@ -84,7 +84,7 @@ public class AvtaleInnhold {
     private String mentorFornavn;
     private String mentorEtternavn;
     private String mentorOppgaver;
-    private Integer mentorAntallTimer;
+    private Double mentorAntallTimer;
     private Integer mentorTimelonn;
     private String mentorTlf;
